rename : 일정시간 들은 음악 도메인 이름 변경
