+++ src/main/java/com/smartcar/dororok/musiclisteninglog/repository/MusicListeningLogRepository.java
@@ -1,7 +1,7 @@
-package com.smartcar.dororok.music.repository;
+package com.smartcar.dororok.musiclisteninglog.repository;
 
 import com.smartcar.dororok.member.domain.entitiy.Member;
-import com.smartcar.dororok.music.domain.Music;
+import com.smartcar.dororok.musiclisteninglog.domain.MusicListeningLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -9,11 +9,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 
-public interface MusicRepository extends JpaRepository<Music, Long> {
+public interface MusicListeningLogRepository extends JpaRepository<MusicListeningLog, Long> {
 
     @Modifying
     @Transactional
-    @Query("DELETE FROM Music m WHERE m.createdAt < :threshold")
+    @Query("DELETE FROM MusicListeningLog m WHERE m.createdAt < :threshold")
     void deleteOlderThan(LocalDateTime threshold);
 
     void deleteByMember(Member member);
